mcp-module/requirements/base.txt
# Core Dependencies
pydantic>=2.0.0
httpx>=0.25.0
websockets>=12.0
redis>=5.0.0

# MCP Framework
fastmcp>=0.1.0
mcp>=1.0.0

# HTTP Server Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0

# LLM Integration
langchain>=0.1.0
transformers>=4.36.0

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
structlog>=24.0.0

# Async Support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
